+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ------------------
 
   * fixed bug in `stubbles\lang\exception\Exception::__toString()` always reporting this class instead of the real exception class
+  * all properties in `stubbles\lang\Properties` where key ends with `password` are automatically stored as `stubbles\lang\SecureString`
 
 
 4.1.1 (2014-08-09)